chore: split eth_call execution.rs into focused modules
